arreglo arquitectura
