--- android/app/src/main/java/com/tournity/Bloc/Controllers/PlayerController.java
@@ -1,4 +1,0 @@
-package com.tournity.Bloc.Controllers;
-
-public class PlayerController {
-}
