--- android/app/src/main/java/com/tournity/Bloc/Controllers/TournamentParticipantsController.java
@@ -1,4 +1,0 @@
-package com.tournity.Bloc.Controllers;
-
-public class TournamentParticipantsController {
-}
