--- android/app/src/main/java/com/tournity/Entities/GroupEntity.java
@@ -1,4 +1,0 @@
-package com.tournity.Entities;
-
-public class GroupEntity {
-}
