--- android/app/src/main/java/com/tournity/Repository/Models/PlayerModel.java
@@ -1,4 +1,0 @@
-package com.tournity.Repository.Models;
-
-public class PlayerModel {
-}
