--- android/app/src/main/java/com/tournity/Repository/Models/TeamModel.java
@@ -1,6 +1,0 @@
-package com.tournity.Repository.Models;
-
-public class TeamModel {
-
-
-}
