--- android/app/src/main/java/com/tournity/Repository/Models/TournamentParticipantsModel.java
@@ -1,4 +1,0 @@
-package com.tournity.Repository.Models;
-
-public class TournamentParticipantsModel {
-}
