--- android/app/src/main/java/com/tournity/Repository/Repositories/PlayerRepository.java
@@ -1,4 +1,0 @@
-package com.tournity.Repository.Repositories;
-
-public class PlayerRepository {
-}
